+++ CG2019112315蔡欣运/CGRenderContext.h
@@ -2,6 +2,9 @@
 #define _CGRENDERCONTEXT_H_INCLUDED
 #include "Vector2.h"
 #include "CGObject.h"
+#include <unordered_map>
+#include <algorithm>
+
 CG_NAMESPACE_ENTER
 class CGView;//实验中视图类的基类，如果类的声明中只需要对象指针，则考虑类预声明。
 class CGRenderContext : public CGObject
@@ -58,22 +61,29 @@ public:
 	//圆绘制算法
 	virtual void MidPointCircle(const Vec2i& center, int radius, unsigned long color);
 	virtual void BresenhamCircle(const Vec2i& center, int radius, unsigned long color);
-	/*
-	virtual void CircleArc(const Vec2i& center, int radius, unsigned long color, float startAngle, float sweepAngle); //逆时针
+	
+	// virtual void CircleArc(const Vec2i& center, int radius, unsigned long color, float startAngle, float sweepAngle); //逆时针
 	
 	//椭圆绘制算法（轴对齐椭圆）
-	virtual void MidPointEllipse(const Vec2i& center, int a, int b, unsigned long color);
-	virtual void EllipseArc(const Vec2i& center, int a, int b, unsigned long color, float startAngle, float sweepAngle); //逆时针
+	//virtual void MidPointEllipse(const Vec2i& center, int a, int b, unsigned long color);
+	//virtual void EllipseArc(const Vec2i& center, int a, int b, unsigned long color, float startAngle, float sweepAngle); //逆时针
 	//有效边表多边形扫描线填充算法
 	virtual void ScanLinePolygonFill(const Vec2iArray& pnts, unsigned long fillcolor);
+	void DrawPixel(int x, int y, unsigned long fillcolor);
+	bool IsOdd(int num);
+	void CalculateBounds(const Vec2iArray & pnts, int & yLow, int & yHigh, int & xLow, int & xHigh);
+	void PrepareMap(std::unordered_map<int, std::vector<float>>& yHitsMap, const Vec2iArray & pnts, int yHigh, int yLow);
+	void GetHits(Vec2i p1, Vec2i p2, int yLow, int yHigh, std::unordered_map<int, std::vector<float>>& yHitsMap);
+	//int Compare(const void * a, const void * b);
 	//边界表示的种子填充算法
-	virtual void BoundFill4(int x, int y, unsigned long boundcolor, unsigned long fillcolor);
-	virtual void BoundFill8(int x, int y, unsigned long boundcolor, unsigned long fillcolor);
+	//virtual void BoundFill4(int x, int y, unsigned long boundcolor, unsigned long fillcolor);
+	//virtual void BoundFill8(int x, int y, unsigned long boundcolor, unsigned long fillcolor);
 	//内点表示的种子填充算法
-	virtual void FloodFill4(int x, int y, unsigned long innercolor, unsigned long fillcolor);
-	virtual void FloodFill8(int x, int y, unsigned long innercolor, unsigned long fillcolor);
+	//virtual void FloodFill4(int x, int y, unsigned long innercolor, unsigned long fillcolor);
+	//virtual void FloodFill8(int x, int y, unsigned long innercolor, unsigned long fillcolor);
 	//扫描线种子填充算法
-	virtual void ScanLineSeedFill(int x, int y, unsigned long boundcolor, unsigned long fillcolor);*/
+	//virtual void ScanLineSeedFill(int x, int y, unsigned long boundcolor, unsigned long fillcolor);
+	//virtual void ScanLinePolygonFill(const Vec2iArray& pnts, unsigned long fillcolor);
 protected:
 	CGView* mView; //通过视图类可以访问程序主框架中的一些设置，比如当前线型、线宽等。
 	HWND mHWND; //通过mView设置
